Move temporary container factory methods
