+++ ref/Meziantou.Framework.TemporaryContainers.cs
@@ -53,10 +53,18 @@ namespace Meziantou.Framework.TemporaryContainers
         public ContainerDefinition(Meziantou.Framework.TemporaryContainers.ImageSource image) { }
         public ContainerDefinition(Meziantou.Framework.TemporaryContainers.ContainerDefinition other) { }
         public virtual Meziantou.Framework.TemporaryContainers.TemporaryContainer CreateContainer() => throw null;
-        public static Meziantou.Framework.TemporaryContainers.RedisContainerDefinition CreateRedis() => throw null;
-        public static Meziantou.Framework.TemporaryContainers.RedisContainerDefinition CreateRedis(Meziantou.Framework.TemporaryContainers.ImageSource image) => throw null;
+    }
+
+    public static class ContainerDefinitionPostgreSqlExtensions
+    {
         public static Meziantou.Framework.TemporaryContainers.PostgreSqlContainerDefinition CreatePostgreSql() => throw null;
         public static Meziantou.Framework.TemporaryContainers.PostgreSqlContainerDefinition CreatePostgreSql(Meziantou.Framework.TemporaryContainers.ImageSource image) => throw null;
+    }
+
+    public static class ContainerDefinitionRedisExtensions
+    {
+        public static Meziantou.Framework.TemporaryContainers.RedisContainerDefinition CreateRedis() => throw null;
+        public static Meziantou.Framework.TemporaryContainers.RedisContainerDefinition CreateRedis(Meziantou.Framework.TemporaryContainers.ImageSource image) => throw null;
     }
 
     public sealed class ContainerEnvironmentCollection : System.Collections.Generic.IEnumerable<System.Collections.Generic.KeyValuePair<string, string>>, System.Collections.IEnumerable
